src/main/java/ru/job4j/cas/cache/Cache.java
package ru.job4j.cas.cache;

import java.util.HashMap;
import java.util.Map;
import java.util.concurrent.ConcurrentHashMap;
import java.util.function.BiFunction;

public class Cache {
    private final Map<Integer, Base> memory = new ConcurrentHashMap<>();

    public boolean add(Base model) {
        return memory.putIfAbsent(model.getId(), model) == null;
    }

    public boolean update(Base model) {
        BiFunction<Integer, Base, Base> upd = (key, base) -> {
            if (base.getVersion() != model.getVersion()) {
                throw new OptimisticException("Versions are not equal");
            }
            Base bs = new Base(key, 1 + base.getVersion());
            bs.setName(model.getName());
            return bs;
        };
        return memory.computeIfPresent(model.getId(), upd) != null;
    }

    public void delete(Base model) {
        memory.remove(model.getId());
    }

    public Map<Integer, Base> getAll() {
        return new HashMap<>(memory);
    }
}